factor out LookupDecoder
